+++ assets/js/just-the-docs.js
@@ -31,10 +31,7 @@ function initNav() {
     }
     if (target) {
       e.preventDefault();
-      const active = target.parentNode.classList.toggle('active');
-      const passive = target.parentNode.classList.toggle('passive');
-      if (active && passive) target.parentNode.classList.toggle('passive');
-      target.ariaPressed = active;
+      target.ariaPressed = target.parentNode.classList.toggle('active');
     }
   });
 
@@ -498,8 +495,7 @@ function scrollNav() {
 }
 
 // Find the nav-list-link that refers to the current page
-// then make it and all enclosing nav-list-item elements active,
-// and make all other folded collections passive
+// then make it and all enclosing nav-list-item elements active.
 
 function activateNav() {
   var target = navLink();
@@ -513,17 +509,6 @@ function activateNav() {
     if (target) {
       target.classList.toggle('active', true);
       target = target.parentNode;
-    }
-  }
-  const elements = document.getElementsByClassName("nav-category-list");
-  for (const element of elements) {
-    const item = element.children[0];
-    const active = item.classList.toggle('active');
-    if (active) {
-      item.classList.toggle('active', false);
-      item.classList.toggle('passive', true);
-    } else {
-      item.classList.toggle('active', true);
     }
   }
 }