+++ CHANGELOG.md
@@ -17,8 +17,10 @@ This website is built from the `HEAD` of the `main` branch of the theme reposito
 
 Code changes to `main` that are *not* in the latest release:
 
+- Remove: unused `.passive` toggle in navigation by [@pdmosses] in [#1335]
 - Fixed: Windows emoji font fallback by [@flanakin] in [#1337]
 
+[#1335]: https://github.com/just-the-docs/just-the-docs/pull/1335
 [#1337]: https://github.com/just-the-docs/just-the-docs/pull/1337
 
 ### New Contributors